+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - IssuingToken resource
 - IssuingTokenRequest resource
 - IssuingTokenDesign resource
+- IssuingTokenActivation resource
 ### Fixed
 - IssuingProduct fundingType, holderType and code were exposed as Settlement, Category and Client; now FundingType, HolderType, Code, plus Client and Settlement from the API